@@ -1,21 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ACMESharpCore" Version="2.0.0.49-beta1" />
+    <PackageReference Include="ACMESharpCore" Version="2.1.0.113" />
     <PackageReference Include="Docker.DotNet" Version="3.125.2" />
     <PackageReference Include="KdSoft.FlatBuffers" Version="1.9.3" />
-    <PackageReference Include="Nito.AsyncEx" Version="5.0.0-pre-05" />
-    <PackageReference Include="RBush" Version="2.0.31" />
+    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.5" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.5" />
+    <PackageReference Include="Nito.AsyncEx" Version="5.0.0" />
+    <PackageReference Include="RBush" Version="2.0.41" />
     <PackageReference Include="System.Numerics.Vectors" Version="4.5.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Discord.Net" Version="2.0.0-beta2-01039" />
-    <PackageReference Include="DnsClient" Version="1.1.1" />
-    <PackageReference Include="Makaretu.Dns.Unicast" Version="0.5.0" />
+    <PackageReference Include="Discord.Net" Version="2.2.0" />
+    <PackageReference Include="DnsClient" Version="1.3.2" />
+    <PackageReference Include="Makaretu.Dns.Unicast" Version="0.11.1" />
   </ItemGroup>
 
   <ItemGroup>
