@@ -2,18 +2,18 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="ConsoleTableExt" Version="2.0.1" />
     <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="2.2.4" />
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
+    <PackageReference Include="ConsoleTableExt" Version="2.0.2" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="TiledSharp" Version="1.0.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Game.Robots\Game.Robots.csproj" />
-    <PackageReference Include="Google.Cloud.Firestore" Version="1.0.0-beta19" />
+    <PackageReference Include="Google.Cloud.Firestore" Version="2.1.0" />
   </ItemGroup>
   <ItemGroup>
     <None Update="Dockerfile">
