@@ -12,6 +12,7 @@
     using Newtonsoft.Json;
     using System;
     using System.Net.Http;
+    using Microsoft.AspNetCore.Mvc.NewtonsoftJson;
 
     public class Startup
     {
@@ -26,7 +27,7 @@
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             services.AddScoped<ISecurityContext, TokenSecurityContext>();
             services.AddMvc().
-                AddJsonOptions(options =>
+                AddNewtonsoftJson(options =>
                 {
                     options.SerializerSettings.Formatting = Newtonsoft.Json.Formatting.Indented;
                     options.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Serialize;
